refactor(α): lint-fence + finish wire migration for upload/auth
Second slice of phase α. Adds the structural guard so the wire-type
contract sticks, and finishes migrating the remaining `Backend*`
declarations across `src/`.

Lint guardrail
* `no-restricted-syntax` rule in `eslint.config.js` rejects any
  `TSTypeAliasDeclaration` / `TSInterfaceDeclaration` named `Backend*`
  inside `src/`. Forces feature code to import from
  `@/shared/types/wire`. Worker code (`worker/**`) is unaffected so the
  wire types stay where they're authored.

Wire migration
* `worker/types/wire/upload.ts` — `BackendUploadRecord` (snake_case wire
  shape used by `MatterFilesPanel.tsx` and the uploads API).
* `worker/types/wire/auth.ts` — `BackendSession` (snake_case session
  record from the better-auth backend).
* `src/shared/lib/uploadsApi.ts` — `BackendUploadRecord` migrated to
  re-export from wire. `BackendUploadResult` and `BackendUploadOptions`
  were misnamed (camelCase, frontend-only API method types — not wire
  shapes); renamed to `UploadResult` / `UploadOptions`.
* `src/shared/types/user.ts` — `BackendSession` migrated to wire.
  `BackendSessionUser` was a misnamed alias for the better-auth library's
  own `BetterAuthSessionUser` shape (not an HTTP wire type); renamed to
  `SessionUser` and consumers in `SessionContext.tsx` and `authClient.ts`
  updated.

Audit fix follow-up
* `package.json` overrides change: `brace-expansion: "^5.0.5"` broke
  ESLint's bundled minimatch (which calls into `expand` as a function —
  an API that disappeared after 4.x). Switched to the nested form
  `minimatch@9.x: { brace-expansion: "^3.0.2" }` so the
  `@typescript-eslint/typescript-estree` chain gets the patched 3.x
  while ESLint's older `minimatch@3.x` chain keeps `brace-expansion@1.x`
  (already `1.1.14`, outside the audit advisory range). Audit still
  reports 0 vulnerabilities.

Other
* `useWorkspaceAutoNavigation.tsx` — added `hasAutoNavigatedRef` and
  `navigationInitiatedRef` to the effect deps that ESLint flagged after
  the lint config rebuild.

Verification
* `npx tsc --noEmit`: 0 errors
* `npx eslint src/`: 0 errors (8 pre-existing warnings)
* `npm audit`: 0 vulnerabilities
* `npm run test:unit`: 243/243 pass
* `npm run test:component`: 75/75 pass
