+++ package.json
@@ -118,6 +118,6 @@
   "overrides": {
     "serialize-javascript": "^7.0.5",
     "undici": "^7.25.0",
-    "brace-expansion": "^5.0.5"
+    "minimatch@9.x": { "brace-expansion": "^3.0.2" }
   }
 }
